requirements.txt
absl-py==0.9.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
django-environ==0.4.5
environ==1.0
future==0.18.2
google-auth==1.20.0
google-auth-oauthlib==0.4.1
grpcio==1.30.0
grpcio-tools==1.30.0
idna==2.10
importlib-metadata==1.7.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.3.0.36
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tb-nightly==2.4.0a20200730
tensorboard-plugin-wit==1.7.0
torch==1.6.0
tqdm==4.48.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
